requirements.txt
Django
djangorestframework
pymysql
gunicorn

